Fix big with files having missing pieces
Was caused the the get random piece function not accounting for the pieces the other peer actually had, so it would request pieces the other peer did not have and get send a blank piece.

I changed some other random bs trying to fix this, but its mostly for the better so ima leave it. Changed:
- Bitfields send at bytes instead of ASCI
- Added file locks